rework the benchmarks
